Lots of updates on this one
Set the response to updates of FO's to be it's own page with good/failed page.

changed behavior of "New" click on FO

redirect new fo to go to dataent.js

dataent add line break for response

added some further fine tuning for visuals

changed foprint to link to foentry as it is the same function

fixed error with resetting table when clicking new.